refactor: TT-401 fixtures changed
