+++ V2/tests/conftest.py
@@ -1,3 +1,3 @@
 # flake8: noqa
-from fixtures import _activity_factory, _create_fake_database
+from fixtures import _activity_factory, _create_fake_database, _insert_activity
 from fixtures import _time_entry_factory
