+++ V2/tests/integration/daos/activities_sql_dao_test.py
@@ -7,14 +7,6 @@ import time_tracker.activities._infrastructure as infrastructure
 from time_tracker._infrastructure import DB
 
 
-@pytest.fixture(name='insert_activity')
-def _insert_activity() -> domain.Activity:
-    def _new_activity(activity: domain.Activity, dao: domain.ActivitiesDao):
-        new_activity = dao.create(activity)
-        return new_activity
-    return _new_activity
-
-
 @pytest.fixture(name='create_fake_dao')
 def _create_fake_dao() -> domain.ActivitiesDao:
     db_fake = DB('sqlite:///:memory:')
@@ -48,7 +40,7 @@ def test_update__returns_an_update_activity__when_an_activity_matching_its_id_is
 ):
     dao = create_fake_dao
     existent_activity = activity_factory()
-    inserted_activity = insert_activity(existent_activity, dao)
+    inserted_activity = insert_activity(existent_activity, dao.db)
 
     expected_description = Faker().sentence()
     updated_activity = dao.update(inserted_activity.id, None, expected_description, None, None)
@@ -75,8 +67,8 @@ def test__get_all__returns_a_list_of_activity_dto_objects__when_one_or_more_acti
     dao = create_fake_dao
     existent_activities = [activity_factory(), activity_factory()]
     inserted_activities = [
-        insert_activity(existent_activities[0], dao),
-        insert_activity(existent_activities[1], dao)
+        insert_activity(existent_activities[0], dao.db),
+        insert_activity(existent_activities[1], dao.db)
     ]
 
     activities = dao.get_all()
@@ -90,7 +82,7 @@ def test_get_by_id__returns_an_activity_dto__when_found_one_activity_that_matche
 ):
     dao = create_fake_dao
     existent_activity = activity_factory()
-    inserted_activity = insert_activity(existent_activity, dao)
+    inserted_activity = insert_activity(existent_activity, dao.db)
 
     activity = dao.get_by_id(inserted_activity.id)
 
@@ -124,7 +116,7 @@ def test_delete__returns_an_activity_with_inactive_status__when_an_activity_matc
 ):
     dao = create_fake_dao
     existent_activity = activity_factory()
-    inserted_activity = insert_activity(existent_activity, dao)
+    inserted_activity = insert_activity(existent_activity, dao.db)
 
     activity = dao.delete(inserted_activity.id)
