+++ V2/time_tracker/time_entries/_infrastructure/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa
-from ._data_persistence import TimeEntriesJsonDao
+from ._data_persistence import TimeEntriesSQLDao
