+++ V2/time_tracker/time_entries/_infrastructure/_data_persistence/__init__.py
@@ -1,2 +1,2 @@
 # flake8: noqa
-from ._time_entries_dao import TimeEntriesJsonDao
+from ._time_entries_dao import TimeEntriesSQLDao
